+++ fairmlhealth/__preprocessing.py
@@ -1,7 +1,6 @@
 '''
 
 '''
-from importlib.util import find_spec
 import numpy as np
 import pandas as pd
 from . import __validation as valid
@@ -9,7 +8,24 @@ from .__validation import ValidationError
 
 
 
-def prep_X(data):
+def analytical_labels(pred_type: str = "binary"):
+    """ Returns a dictionary of category labels used by analytical functions
+    Args:
+        pred_type (b): number of classes in the prediction problem
+    """
+    valid_pred_types = ["binary", "multiclass", "regression"]
+    if pred_type not in valid_pred_types:
+        raise ValueError(f"pred_type must be one of {valid_pred_types}")
+    c_note = "" if pred_type == "binary" else " (Weighted Avg)"
+    lbls = {'gf_label': "Group Fairness",
+            'if_label': "Individual Fairness",
+            'mp_label': f"Model Performance{c_note}",
+            'dt_label': "Data Metrics"
+            }
+    return lbls
+
+
+def prep_data(data):
     """ Ensures that data are in the correct format
 
     Returns:
@@ -19,13 +35,17 @@ def prep_X(data):
         if isinstance(data, pd.Series):
             X = pd.DataFrame(data, columns=[data.name])
         else:
-            X = pd.DataFrame(data, columns=['X'])
+            X = pd.DataFrame(data)
+        # Ensure that some unique identifier is present for each/any column(s)
+        if not any(X.columns):
+            X.columns = [str(i) for i in range(len(X.columns))]
     else:
-        X = data
-    # Convert columns to numeric type if they do not contain strings
+        X = data.copy(deep=True)
+    # Convert columns that do not contain any strings to numeric type
     for col in X.columns:
         X.loc[:, col] = pd.to_numeric(X[col], errors='ignore')
     return X
+
 
 def prep_prtc_attr(arr):
     if not isinstance(arr, pd.DataFrame):
@@ -35,75 +55,45 @@ def prep_prtc_attr(arr):
             pa_name = 'protected_attribute'
             prtc_attr = pd.DataFrame(arr, columns=[pa_name])
     else:
-        prtc_attr = arr
+        prtc_attr = arr.copy(deep=True)
     prtc_attr.reset_index(inplace=True, drop=True)
     return prtc_attr
 
+
 def prep_targets(arr, prtc_attr=None):
     if isinstance(arr, (np.ndarray, pd.Series)):
-        y_true = pd.DataFrame(arr)
+        targets = pd.DataFrame(arr)
     else:
-        y_true = arr
+        targets = arr.copy(deep=True)
     if prtc_attr is not None:
-        y_true = pd.concat([prtc_attr, y_true.reset_index(drop=True)], axis=1)
-        y_true.set_index(prtc_attr.columns.tolist(), inplace=True)
-    return y_true
+        targets = pd.concat([prtc_attr, targets.reset_index(drop=True)], axis=1)
+        targets.set_index(prtc_attr.columns.tolist(), inplace=True)
+    else:
+        targets.reset_index(drop=True, inplace=True)
+    return targets
 
-def prep_preds(arr, y_col=None, prtc_attr=None):
+
+def prep_preds(arr, y_col=None, prtc_attr=None, name="predictions"):
     if isinstance(arr, np.ndarray):
-        y_pred = pd.DataFrame(arr)
+        preds = pd.DataFrame(arr)
     else:
-        y_pred = arr
+        preds = arr.copy(deep=True)
     if prtc_attr is not None:
-        y_pred = pd.concat([prtc_attr, y_pred.reset_index(drop=True)], axis=1)
-        y_pred.set_index(prtc_attr.columns.tolist(), inplace=True)
+        preds = pd.concat([prtc_attr, preds.reset_index(drop=True)], axis=1)
+        preds.set_index(prtc_attr.columns.tolist(), inplace=True)
     else:
-        pass
+        preds.reset_index(drop=True, inplace=True)
     if y_col is None:
         raise ValidationError(
-            "Cannot evaluate predictions without ground truth")
+            f"Cannot evaluate {name} without ground truth")
     else:
-        y_pred.columns = y_col
-    return y_pred
-
-
-def prep_probs(arr, y_col=None, prtc_attr=None):
-    if isinstance(arr, np.ndarray):
-        y_prob = pd.DataFrame(arr)
-    else:
-        y_prob = arr
-    if prtc_attr is not None:
-        y_prob = pd.concat([prtc_attr, y_prob.reset_index(drop=True)], axis=1)
-        y_prob.set_index(prtc_attr.columns.tolist(), inplace=True)
-    else:
-        pass
-    if y_col is None:
-        raise ValidationError(
-            "Cannot evaluate probabilities without ground truth")
-    else:
-        y_prob.columns = y_col
-    return y_prob
-
-def report_labels(pred_type: str = "binary"):
-    """ Returns a dictionary of category labels used by reporting functions
-    Args:
-        pred_type (b): number of classes in the prediction problem
-    """
-    valid_pred_types = ["binary", "multiclass", "regression"]
-    if pred_type not in valid_pred_types:
-        raise ValueError(f"pred_type must be one of {valid_pred_types}")
-    c_note = "" if pred_type == "binary" else " (Weighted Avg)"
-    report_labels = {'gf_label': "Group Fairness",
-                     'if_label': "Individual Fairness",
-                     'mp_label': f"Model Performance{c_note}",
-                     'dt_label': "Data Metrics"
-                     }
-    return report_labels
+        preds.columns = [y_col]
+    return preds
 
 
 def standard_preprocess(X, prtc_attr=None, y_true=None, y_pred=None,
                         y_prob=None, priv_grp=1):
-    """ Formats data for use by fairness reporting functions.
+    """ Formats data for use by fairness analytical functions.
     Args:
         X (array-like): Sample features
         prtc_attr (named array-like): values for the protected attribute
@@ -117,22 +107,26 @@ def standard_preprocess(X, prtc_attr=None, y_true=None, y_pred=None,
     Returns:
         Tuple containing formatted versions of all passed args.
     """
-    valid.validate_report_input(X, y_true, y_pred, y_prob, prtc_attr, priv_grp)
+    valid.validate_analytical_input(X, y_true, y_pred, y_prob, prtc_attr, priv_grp)
 
     # Format inputs to required datatypes
-    X = prep_X(X)
+    X = prep_data(X)
     # Format protected attributes
     if prtc_attr is not None:
         prtc_attr = prep_prtc_attr(prtc_attr)
     #
-    y_col = None
+    _y = None
     if y_true is not None:
         y_true = prep_targets(y_true, prtc_attr)
-        y_col = y_true.columns
+        if not any(y_true.columns) or y_true.columns[0] == 0:
+            _y = y_cols()['col_names']['yt']
+            y_true.columns = [_y]
+        else:
+            _y = y_true.columns[0]
     if y_pred is not None:
-        y_pred = prep_preds(y_pred, y_col, prtc_attr)
+        y_pred = prep_preds(y_pred, _y, prtc_attr, name="predictions")
     if y_prob is not None:
-        y_prob = prep_probs(y_prob, y_col, prtc_attr)
+        y_prob = prep_preds(y_prob, _y, prtc_attr, name="probabilities")
     #
     return (X, prtc_attr, y_true, y_pred, y_prob)
 
@@ -140,7 +134,7 @@ def standard_preprocess(X, prtc_attr=None, y_true=None, y_pred=None,
 def stratified_preprocess(X, y_true=None, y_pred=None, y_prob=None,
                           features:list=None):
     """
-    Runs validation and formats data for use in stratified reports
+    Runs validation and formats data for use in stratified tables
 
     Args:
         df (pandas dataframe or compatible object): sample data to be assessed
@@ -154,13 +148,13 @@ def stratified_preprocess(X, y_true=None, y_pred=None, y_prob=None,
     Requirements:
         - Each feature must be discrete to run stratified analysis, and must be
         binary to run the assessment. If any data are not discrete and there
-        are more than 11 values, the reporter will reformat those data into
+        are more than 11 values, the tool will reformat those data into
         quantiles
     """
     #
     X, _, y_true, y_pred, y_prob = \
         standard_preprocess(X, prtc_attr=None, y_true=y_true, y_pred=y_pred,
-                           y_prob=y_prob)
+                            y_prob=y_prob)
     # Attach y variables and subset to expected columns
     yt, yh, yp = y_cols()['col_names'].values()
     df = X.copy()
@@ -186,7 +180,7 @@ def stratified_preprocess(X, y_true=None, y_pred=None, y_prob=None,
 
 def y_cols(df=None):
     ''' Returns a dict of hidden column names for each
-        of the y values used in stratified reporting functions, the keys for
+        of the y values used in stratified table functions, the keys for
         which are as follows: "yt"="y true"; "yh"="y predicted";
         "yp"="y probabilities". This allows for consistent references that are
         not likely to be found among the actual columns of the data (e.g., so
@@ -199,9 +193,9 @@ def y_cols(df=None):
                 of known names; names that are not found will be dropped from
                 the results. Defaults to None.
     '''
-    y_names = {'col_names': {'yt': '__fairmlhealth_y_true',
-                            'yh': '__fairmlhealth_y_pred',
-                            'yp': '__fairmlhealth_y_prob'},
+    y_names = {'col_names': {'yt': '__y_true',
+                            'yh': '__y_pred',
+                            'yp': '__y_prob'},
               'disp_names': {'yt': 'Target',
                              'yh': 'Pred.',
                              'yp': 'Prob.'}