+++ test/test_docs.py
@@ -1,4 +1,4 @@
-''' 
+'''
 Test script to flag obvious errors in markdown documentation files. Example
 errors included broken http(s) links. At least verifies the presence of
 documentation that is expected.
@@ -6,7 +6,7 @@ documentation that is expected.
 import os
 from pathlib import Path
 import pytest
-from .__utils import (get_urls, get_url_status,  is_test_environment,
+from .__test__utils import (get_urls, get_url_status,  is_test_environment,
                       is_url_valid, URLError)
 
 
@@ -99,5 +99,5 @@ def test_templatesREADME():
 
 def test_tutorialsREADME():
     repo_main = base_dir()
-    file = os.path.join(repo_main, "tutorials_and_examples", "README.md")
+    file = os.path.join(repo_main, "examples_and_tutorials", "README.md")
     validate_markdown(file)