+++ test/test_example_toolusage.py
@@ -11,7 +11,7 @@ import warnings
 def test_example_toolusage():
     this_dir = os.path.dirname(os.path.abspath(__file__))
     nb_name = "Example-ToolUsage.ipynb"
-    nb_path = os.path.join(this_dir, "..", "tutorials_and_examples", nb_name)
+    nb_path = os.path.join(this_dir, "..", "examples_and_tutorials", nb_name)
     nb, err = validate_notebook(nb_path)
 
     if any(err):